Moved factory to own namespace
