+++ tests/Unit/Factories/CssInlinerFactoryTest.php
@@ -2,12 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Pelago\Emogrifer\Tests\Unit;
+namespace Pelago\Emogrifer\Tests\Unit\Factories;
 
 use DOMDocument;
 use Pelago\Emogrifier\CssInliner;
-use Pelago\Emogrifier\CssInlinerFactory;
+use Pelago\Emogrifier\Factories\CssInlinerFactory;
 use PHPUnit\Framework\TestCase;
+use Pelago\Emogrifer\Tests\Unit\CssInlinerTest;
 
 /**
  * Test case.
